fixed direction for moving to new rooms, gel movement, sped up player, bounds for bomb and sword explosion, dragon shooting backward, etc

+++ LoZGame/Players/PlayerClasses/PlayerEssentials.cs
@@ -15,7 +15,7 @@
 
         public Color CurrentTint { get; set; }
 
-        public int MoveSpeed { get; set; }
+        public float MoveSpeed { get; set; }
 
         public int DamageTimer { get; set; }
